Avoid duplication & fix code smell
